Run ucc with preferences directory
This allows the server to separate base game content (in /ut-server)
from user-modified configuration files and mods (which will now live in
/.loki/ut)

Additionally, the UT server will now run as a non-root user ("ut") and
the configuration file will no longer be fiddled with on startup. The
only env vars that are now read by the startup script are:

 - STARTUP_MAP
 - GAME_MODE
 - MUTATORS

Any other configuration should be done manually against
UnrealTournament.ini etc. in the /.loki/ut volume.